fmt code

+++ foundry/counter/src/CallUsage.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.20;
 
 contract Receiver {
-    event Received(address caller, uint amount, string message);
+    event Received(address caller, uint256 amount, string message);
 
     /*
         Which function is called, fallback() or receive()?
@@ -24,20 +24,18 @@ contract Receiver {
         emit Received(msg.sender, msg.value, "Fallback was called");
     }
 
-    function foo(string memory _message, uint _x) public payable returns (uint) {
+    function foo(string memory _message, uint256 _x) public payable returns (uint256) {
         emit Received(msg.sender, msg.value, _message);
 
         return _x + 1;
     }
 
     // receiver must be empty
-    receive() external payable {
-    }
+    receive() external payable {}
 
-    function getBalance() public view returns (uint) {
+    function getBalance() public view returns (uint256) {
         return address(this).balance;
     }
-
 }
 
 contract Caller {
@@ -47,20 +45,16 @@ contract Caller {
     // contract Receiver, but we do know the address of contract Receiver and the function to call.
     function testCallFoo(address payable _addr) public payable {
         // You can send ether and specify a custom gas amount
-        (bool success, bytes memory data) = _addr.call{value: msg.value, gas: 5000}(
-            abi.encodeWithSignature("foo(string,uint256)", "call foo", 123)
-        );
+        (bool success, bytes memory data) =
+            _addr.call{value: msg.value, gas: 5000}(abi.encodeWithSignature("foo(string,uint256)", "call foo", 123));
 
         emit Response(success, data);
     }
 
     // Calling a function that does not exist triggers the fallback function.
     function testCallDoesNotExist(address payable _addr) public payable {
-        (bool success, bytes memory data) = _addr.call{value: msg.value}(
-            abi.encodeWithSignature("doesNotExist()")
-        );
+        (bool success, bytes memory data) = _addr.call{value: msg.value}(abi.encodeWithSignature("doesNotExist()"));
 
         emit Response(success, data);
     }
 }
-
